# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Disney API",
    "version": "1.0.0"
  },
  "schemes": [
    "https",
    "http"
  ],
  "produces": [
    "application/json"
  ],
  "consumes": [
    "application/x-www-form-urlencoded"
  ],
  "securityDefinitions": {
    "api_key": {
      "type": "apiKey",
      "name": "authorization",
      "in": "header"
    }
  },
  "security": [
    {
      "api_key": []
    }
  ],
  "definitions": {
...
    "Genre": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "image": {
          "type": "string"
        },
        "links": {
          "type": "object",
          "properties": {
            "self": {
              "type": "object",
              "properties": {
                "href": {
                  "type": "string"
                }
              }
            },
            "films": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "href": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
    "Character": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "image": {
          "type": "string"
        },
        "weight": {
          "type": "number"
        },
        "age": {
          "type": "number"
        },
        "story": {
          "type": "string"
        },
        "links": {
          "type": "object",
          "properties": {
            "self": {
              "type": "object",
              "properties": {
                "href": {
                  "type": "string"
                }
              }
            },
            "films": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "href": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
    "CharacterList": {
      "type": "array",
      "items": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "image": {
            "type": "string"
          },
          "self": {
            "type": "object",
            "properties": {
              "href": {
                "type": "string"
              }
            }
          }
        }
      }
    },
    "Film": {
      "type": "object",
      "properties": {
        "title": {
          "type": "string"
        },
        "image": {
          "type": "string"
        },
        "releaseDate": {
          "type": "string"
        },
        "rating": {
          "type": "number"
        },
        "genreId": {
          "type": "string"
        },
        "links": {
          "type": "object",
          "properties": {
            "self": {
              "type": "object",
              "properties": {
                "href": {
                  "type": "string"
                }
              }
            },
            "characters": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "href": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
    "FilmList": {
      "type": "array",
      "items": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "image": {
            "type": "string"
          },
          "releaseDate": {
            "type": "string"
          },
          "self": {
            "type": "object",
            "properties": {
              "href": {
                "type": "string"
              }
            }
          }
        }
      }
    }
  },
  "paths": {
...
    "/characters": {
      "get": {
        "tags": [
          "Characters"
        ],
        "summary": "Find all characters",
        "description": "Find all characters with certain parameters, leave blank to get all characters",
        "operationId": "getCharacters",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "weight",
            "in": "query",
            "required": false,
            "type": "number"
          },
          {
            "name": "age",
            "in": "query",
            "required": false,
            "type": "number"
          },
          {
            "name": "filmName",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Returns all characters found characters",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/CharacterList"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Unsuccessful operation",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Characters"
        ],
        "summary": "Add a new character",
        "description": "Add a new character, associated films need to be created first",
        "operationId": "addCharacter",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "name",
            "in": "formData",
            "required": true,
            "type": "string"
          },
          {
            "name": "image",
            "in": "formData",
            "required": false,
            "type": "file"
          },
          {
            "name": "story",
            "in": "formData",
            "required": true,
            "type": "string"
          },
          {
            "name": "weight",
            "in": "formData",
            "required": false,
            "type": "number"
          },
          {
            "name": "age",
            "in": "formData",
            "required": false,
            "type": "number"
          },
          {
            "name": "filmIds",
            "in": "formData",
            "required": false,
            "collectionFormat": "multi",
            "description": "Array of existing associated film ids",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Returns the added character",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "$ref": "#/definitions/Character"
                }
              }
            }
          },
          "400": {
            "description": "Unsuccessful operation",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/movies": {
      "get": {
        "tags": [
          "Movies"
        ],
        "summary": "Find all movies",
        "description": "Find all movies with certain parameters, leave blank to find all movies",
        "operationId": "getMovies",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "title",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "genre",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "order",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Returns all movies",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/FilmList"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Unsuccessful operation",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Movies"
        ],
        "summary": "Add a movie",
        "description": "Add a movie",
        "operationId": "addMovie",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "title",
            "in": "formData",
            "required": true,
            "type": "string"
          },
          {
            "name": "image",
            "in": "formData",
            "required": false,
            "type": "file"
          },
          {
            "name": "releaseDate",
            "in": "formData",
            "required": true,
            "type": "string"
          },
          {
            "name": "rating",
            "in": "formData",
            "required": true,
            "type": "number"
          },
          {
            "name": "genreId",
            "in": "formData",
            "required": false,
            "type": "string"
          },
          {
            "name": "characterIds",
            "in": "formData",
            "required": false,
            "collectionFormat": "multi",
            "description": "Array of existing associated character ids",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Returns the added movie",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "$ref": "#/definitions/Film"
                }
              }
            }
          },
          "400": {
            "description": "Unsuccessful operation",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/movies/{id}": {
      "get": {
        "tags": [
          "Movies"
        ],
        "summary": "Find a movie by id",
        "description": "Find a movie by id",
        "operationId": "getMovie",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Returns the movie",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "$ref": "#/definitions/Film"
                }
              }
            }
          },
          "400": {
            "description": "Unsuccessful operation",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Movies"
        ],
        "summary": "Update a movie",
        "description": "Update a movie",
        "operationId": "updateMovie",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "title",
            "in": "formData",
            "required": false,
            "type": "string"
          },
          {
            "name": "image",
            "in": "formData",
            "required": false,
            "type": "file"
          },
          {
            "name": "releaseDate",
            "in": "formData",
            "required": false,
            "type": "string"
          },
          {
            "name": "rating",
            "in": "formData",
            "required": false,
            "type": "number"
          },
          {
            "name": "genreId",
            "in": "formData",
            "required": false,
            "type": "string"
          },
          {
            "name": "characterIds",
            "in": "formData",
            "required": false,
            "collectionFormat": "multi",
            "description": "Array of existing associated character ids, THIS WILL BE REPLACED ENTIRELY",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns the updated movie",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "$ref": "#/definitions/Film"
                }
              }
            }
          },
          "400": {
            "description": "Unsuccessful operation",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Movies"
        ],
        "summary": "Delete a movie",
        "description": "Delete a movie",
        "operationId": "deleteMovie",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "204": {
            "description": "Successful operation"
          },
          "400": {
            "description": "Unsuccessful operation",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/genres": {
      "get": {
        "tags": [
          "Genres"
        ],
        "summary": "Get all genres",
        "description": "Get all genres",
        "operationId": "getGenres",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Returns all genres",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/Genre"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Genres"
        ],
        "summary": "Add a genre",
        "description": "Add a genre",
        "operationId": "addGenre",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "name",
            "in": "formData",
            "required": true,
            "type": "string"
          },
          {
            "name": "image",
            "in": "formData",
            "required": false,
            "type": "file"
          }
        ],
        "responses": {
          "201": {
            "description": "Returns the added genre",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "$ref": "#/definitions/Genre"
                }
              }
            }
          }
        }
      }
    },
    "/genres/{id}": {
      "get": {
        "tags": [
          "Genres"
        ],
        "summary": "Find a genre by id",
        "description": "Find a genre by id",
        "operationId": "getGenre",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Returns the genre",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "$ref": "#/definitions/Genre"
                }
              }
            }
          },
          "404": {
            "description": "Genre not found",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Genres"
        ],
        "summary": "Update a genre",
        "description": "Update a genre",
        "operationId": "updateGenre",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "name",
            "in": "formData",
            "required": false,
            "type": "string"
          },
          {
            "name": "image",
            "in": "formData",
            "required": false,
            "type": "file"
          }
        ],
        "responses": {
          "200": {
            "description": "Returns the updated genre",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "$ref": "#/definitions/Genre"
                }
              }
            }
          },
          "404": {
            "description": "Genre not found",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Genres"
        ],
        "summary": "Delete a genre",
        "description": "Delete a genre",
        "operationId": "deleteGenre",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "204": {
            "description": "Successful operation"
          },
          "404": {
            "description": "Genre not found",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    }
  }
}